+++ 004_earthquakes_eruptions/01_in_data/earthquake.def
