+++ 004_earthquakes_eruptions/01_in_data/info_esmeraldas.txt
