+++ 004_earthquakes_eruptions/01_in_data/meca_turkeysyria.txt
