+++ 004_earthquakes_eruptions/01_la_palma_eruption.py
