+++ 004_earthquakes_eruptions/02_out_figs/esmeraldas_earthquake.png
