+++ 004_earthquakes_eruptions/02_out_figs/la_palma_eruption.png
