+++ 004_earthquakes_eruptions/02_out_figs/marocco_earthquake.png
