+++ 004_earthquakes_eruptions/02_out_figs/tonga_eruption.png
