+++ 004_earthquakes_eruptions/02_out_figs/turkey_syria_earthquake.png
