--- 004_earthquakes_eruptions/03_esmeraldas/earthquake.def
@@ -1,8 +1,0 @@
-# Self-defined symbol earthquake
-
-# For usage in GMT / PyGMT
-0	0	0.5	c
-0	0	0.4	c	-Gwhite -W0.01p,white
-0	0	0.3	c
-0	0	0.2	c	-Gwhite -W0.01p,white
-0	0	0.1	c
