+++ 004_earthquakes_eruptions/06_japan_earthquake_BFO.py
